Move engine classes into hierarchy structure
