+++ src/audio/engine/constants.js
@@ -1,0 +1,2 @@
+export const AUDIO_CONTEXT = 'audioContext';
+export const DEFAULT_NOTE_LENGTH = 0.5;
