+++ src/audio/engine/scheduler/connector/gain/GainNodeConnector.js
@@ -1,5 +1,5 @@
-import EnvelopeValueScheduler from './EnvelopeValueScheduler';
-import { DEFAULT_NOTE_LENGTH } from './WebAudioEngine';
+import EnvelopeValueScheduler from './envelope/EnvelopeValueScheduler';
+import { DEFAULT_NOTE_LENGTH } from 'audio/engine/constants';
 
 class GainNodeConnector {
