+++ src/audio/engine/scheduler/connector/gain/envelope/EnvelopeValueScheduler.js
@@ -1,4 +1,4 @@
-import { DEFAULT_NOTE_LENGTH } from './WebAudioEngine';
+import { DEFAULT_NOTE_LENGTH } from 'audio/engine/constants';
 
 class EnvelopeValueScheduler {
